@@ -1,12 +1,16 @@
 {
     "name": "@dashlane/cli-documentation",
-    "packageManager": "yarn@4.2.2",
+    "type": "module",
+    "packageManager": "yarn@4.5.1",
     "dependencies": {
-        "next": "^14.2.10",
-        "nextra": "^2.13.4",
-        "nextra-theme-docs": "^2.13.4",
+        "next": "^15.0.3",
+        "nextra": "^3.2.3",
+        "nextra-theme-docs": "^3.2.3",
         "react": "^18.3.1",
         "react-dom": "^18.3.1",
-        "sharp": "^0.33.4"
+        "sharp": "^0.33.5"
+    },
+    "devDependencies": {
+        "@types/react": "18.3.12"
     }
 }
