@@ -55,55 +55,55 @@
     },
     "devDependencies": {
         "@aivenio/tsc-output-parser": "^2.1.1",
-        "@eslint/eslintrc": "^3.1.0",
-        "@eslint/js": "^9.4.0",
+        "@eslint/eslintrc": "^3.2.0",
+        "@eslint/js": "^9.15.0",
         "@types/async": "^3.2.24",
-        "@types/better-sqlite3": "^7.6.10",
-        "@types/chai": "^4.3.16",
+        "@types/better-sqlite3": "^7.6.11",
+        "@types/chai": "^5.0.1",
         "@types/eslint__js": "^8.42.3",
         "@types/inquirer": "^9.0.7",
         "@types/libsodium-wrappers": "^0.7.14",
-        "@types/mocha": "^10.0.6",
-        "@types/node": "^18.19.34",
-        "@yao-pkg/pkg": "^5.12.0",
-        "chai": "^5.1.1",
-        "esbuild": "^0.21.5",
-        "eslint": "^9.4.0",
+        "@types/mocha": "^10.0.9",
+        "@types/node": "^22.9.0",
+        "@yao-pkg/pkg": "^6.1.1",
+        "chai": "^5.1.2",
+        "esbuild": "^0.24.0",
+        "eslint": "^9.15.0",
         "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-prettier": "^5.1.3",
-        "execa": "^9.2.0",
-        "husky": "^9.0.11",
-        "mocha": "^10.4.0",
-        "prettier": "^3.3.1",
+        "eslint-plugin-prettier": "^5.2.1",
+        "execa": "^9.5.1",
+        "husky": "^9.1.7",
+        "mocha": "^10.8.2",
+        "prettier": "^3.3.3",
         "ts-node": "^10.9.2",
-        "typescript": "^5.4.5",
-        "typescript-eslint": "^7.12.0"
+        "typescript": "^5.6.3",
+        "typescript-eslint": "^8.15.0"
     },
     "dependencies": {
         "@dashlane/nsm-attestation": "^1.0.2",
-        "@inquirer/prompts": "^5.0.5",
+        "@inquirer/prompts": "^7.1.0",
         "@json2csv/plainjs": "^7.0.6",
         "@json2csv/transforms": "^7.0.6",
         "@napi-rs/clipboard": "^1.1.2",
         "@napi-rs/keyring": "^1.1.6",
-        "@node-rs/argon2": "^1.8.3",
-        "ajv": "^8.16.0",
+        "@node-rs/argon2": "^2.0.0",
+        "ajv": "^8.17.1",
         "ajv-formats": "^3.0.1",
-        "better-sqlite3": "^11.0.0",
+        "better-sqlite3": "^11.5.0",
         "commander": "^12.1.0",
-        "got": "^14.4.1",
-        "jsonpath-plus": "^10.0.0",
-        "libsodium-wrappers": "^0.7.13",
+        "got": "^14.4.4",
+        "jsonpath-plus": "^10.2.0",
+        "libsodium-wrappers": "^0.7.15",
         "node-mac-auth": "^1.1.0",
         "otplib": "^12.0.1",
-        "playwright-core": "^1.44.1",
-        "winston": "^3.13.0",
+        "playwright-core": "^1.49.0",
+        "winston": "^3.17.0",
         "xml-js": "^1.6.11",
         "zlib": "^1.0.5"
     },
     "engines": {
-        "node": ">=18",
+        "node": ">=22",
         "yarn": ">=4.2.2"
     },
-    "packageManager": "yarn@4.2.2"
+    "packageManager": "yarn@4.5.1"
 }
