@@ -1,17 +1,18 @@
-const withNextra = require('nextra')({
+import nextra from 'nextra';
+
+const withNextra = nextra({
     theme: 'nextra-theme-docs',
-    themeConfig: './theme.config.jsx',
+    themeConfig: './theme.config.jsx'
 });
 
 let assetPrefix = '';
 let basePath = '';
 
-module.exports = {
-    ...withNextra(),
+export default withNextra({
     images: {
         unoptimized: true,
     },
     assetPrefix: assetPrefix,
     basePath: basePath,
     output: 'export'
-};
+});
